% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotatedDMRs2Enrichr.R
\name{annotatedDMRs2Enrichr}
\alias{annotatedDMRs2Enrichr}
\title{Query different databases to find enriched proceses.}
\usage{
annotatedDMRs2Enrichr(
  annotatedDMRs,
  active.features = c("promoters", "heads"),
  stat.filter = "P.value",
  stat.thr = 0.01,
  db = NULL
)
}
\arguments{
\item{annotatedDMRs}{anotated DMRs list resulting from annotateDMRs() or scoreAnnotatedDMRs()}

\item{active.features}{annotation level from the gene symbols are taken. Default = c('promoters', 'heads')}

\item{stat.filter}{character indicating which type of statistics use for filtering results. Accepted values: 'P.value', 'Adjusted.P.value' or 'Overlap'. Default 'P.value'.}

\item{stat.thr}{numeric value indicating the threshold to use for selcted statistical test. Default 0.01.}

\item{db}{vector of characters indicating DBs to query. Default c("ClinVar_2019", "OMIM_Disease", "Elsevier_Pathway_Collection", "MSigDB_Hallmark_2020", "MSigDB_Oncogenic_Signatures", "GO_Biological_Process_2018", "Human_Phenotype_Ontology", "KEGG_2016", "NCI-Nature_2016", "Panther_2016", "Reactome_2016", "WikiPathways_2019_Human")}
}
\value{
data.frame with enriched processes
}
\description{
Query gene symbols from resulting annotations list in order to find enriched proceses.
}
