% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotatedDMRs2Exprs.R
\name{annotatedDMRs2Exprs}
\alias{annotatedDMRs2Exprs}
\title{Compute correlation between expression and methylation levels.}
\usage{
annotatedDMRs2Exprs(
  annotatedDMRs,
  expressionProfile,
  active.features = c("promoters", "heads"),
  col.genes = 0,
  col.stat = 6,
  stat.thr = 0.05,
  col.logFC = 2,
  logfc.thr = 0,
  convert.genes = FALSE,
  convert.from,
  beta.thr = 0.3,
  overlap.param.thr = 100,
  param.type = "overlap.length",
  line.col = "lightgrey",
  lmfit.col1 = "red",
  lmfit.col2 = "green",
  pal = "RdGy",
  plot.type = "splitted",
  show.text = FALSE,
  cor.type = "pearson",
  filter.by.genes = NULL,
  return.table = FALSE
)
}
\arguments{
\item{annotatedDMRs}{anotated DMRs list resultingfrom annotateDMRs() or scoreAnnotatedDMRs()}

\item{expressionProfile}{data.frame, expression profile.}

\item{active.features}{character vectors, containing features to correlate. Must be from names of resulting list from annotateDMRs. Additional feature names can be first exons (exons1) or first intron (intron1). To use more than one feature use c(). Default = c("promoters", "heads")}

\item{col.genes}{numeric, the column of expressionProfile data.frame with gene Ids. If NULL geneIDs will be taken from rownames() of expressionProfile. Default = 0.}

\item{col.stat}{numeric, the column of expressionProfile data.frame with the statistics to use. Default = 6.}

\item{stat.thr}{numeric, threshold for statistical significance. Default = 0.05}

\item{col.logFC}{numeric, the column of expressionProfile data.frame with log. fold change. Default = 2}

\item{logfc.thr}{numeric, threshold value for log. fold change. Default = 0.}

\item{convert.genes}{boolean, used to indicate if gene ids have to be translated in official gene symbols. Default = FALSE}

\item{convert.from}{character, used annotation for gene in expressionProfile to be converted to symbols gene IDs. Accepted: c("ENTREZID" ,"EXONID" ,"GENEBIOTYPE" ,"GENEID" ,"GENENAME" ,"PROTDOMID" ,"PROTEINDOMAINID" ,"PROTEINDOMAINSOURCE" ,"PROTEINID" ,""SEQSTRAND" ,"SYMBOL" ,"TXBIOTYPE" ,"TXID" ,"TXNAME" ,"UNIPROTID")}

\item{beta.thr}{numeric, beta difference threshold value. Default = 0.3.}

\item{overlap.param.thr}{nuemric, threshold value for selected parameter to filter methylations overlapping the selected features. Default = 100}

\item{param.type}{character, threshold parameter to filter methylations overlapping the selected features. Accetped c("dmr.length", "overlap.length", "overlap.percentage"). Default = "overlap.length".}

\item{line.col}{character, color of lines at x=0, y=0. Default = "lightgray"}

\item{lmfit.col1}{character, color of linear model line 1 or for simple plot. Default = "red"}

\item{lmfit.col2}{character, color of linear model line 2. Default = "green"}

\item{pal}{character, color palette. hcl.pals() to show available. Default = "RdGy"}

\item{plot.type}{character, compute or not different linear models for upregulated and downregulated genes. Accepted: "simple" or "splitted". Default = "splitted".}

\item{show.text}{logical, indicating if print gene names in the final plot. Accepted values: TRUE or FALSE. Default = FALSE.}

\item{cor.type}{character, correlation method. Available "pearson", "kendall" or "spearman" correlation, Default = "pearson"}

\item{filter.by.genes}{character vectors, gene symbols used for filtering output}

\item{return.table}{logical, TRUE return a data.frame instead a plot. Default = FALSE}
}
\value{
plot of correlation or (if return.table = TRUE) a data.frame of genes expression associated to beta methylation values.
}
\description{
Compute correlation between expression and methylation levels.
}
