% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotatedEnh2Exprs.R
\name{annotatedEnh2Exprs}
\alias{annotatedEnh2Exprs}
\title{Compute expression methylation correlation}
\usage{
annotatedEnh2Exprs(
  annotatedEnhancers,
  expressionProfile,
  hg = "hg19",
  enhancer.db = "FANTOM5",
  col.genes = 0,
  col.stat = 6,
  stat.thr = 0.05,
  col.logFC = 2,
  logfc.thr = 0.5,
  convert.genes = FALSE,
  convert.from,
  beta.thr = 0.3,
  overlap.param.thr = 40,
  param.type = "overlap.percentage",
  line.col = "lightgrey",
  lmfit.col1 = "red",
  lmfit.col2 = "green",
  pal = "RdGy",
  plot.type = "splitted",
  show.text = FALSE,
  cor.type = "pearson",
  return.table = FALSE
)
}
\arguments{
\item{annotatedEnhancers}{data.frame. It corresponds to resulting list from annotateEnhancers()}

\item{expressionProfile}{expression profile data.frame}

\item{hg}{character, "hg19", "hg38". Version of the enhancer database. Default = "hg19"}

\item{enhancer.db}{character, which database to use between 'FANTOM5' or '4DGenome'. Default = "FANTOM5"}

\item{col.genes}{numeric, the column of expressionProfile data.frame with gene Ids. If NULL geneIDs will be taken from rownames() of expressionProfile. Default = 0.}

\item{col.stat}{numeric, the column of expressionProfile data.frame with the statistics to use. Default = 6.}

\item{stat.thr}{threshold for statistical significance. Default = 0.05}

\item{col.logFC}{numeric, the column of expressionProfile data.frame with log. fold change. Default = 2}

\item{logfc.thr}{numeric, threshold value for log. fold change. Default = 0.}

\item{convert.genes}{logical, used to indicate if gene ids have to be translated in official gene symbols. Default = FALSE}

\item{convert.from}{character, used annotation for gene in expressionProfile to be converted to symbols gene IDs. Accepted: c("ENTREZID" ,"EXONID" ,"GENEBIOTYPE" ,"GENEID" ,"GENENAME" ,"PROTDOMID" ,"PROTEINDOMAINID" ,"PROTEINDOMAINSOURCE" ,"PROTEINID" ,""SEQSTRAND" ,"SYMBOL" ,"TXBIOTYPE" ,"TXID" ,"TXNAME" ,"UNIPROTID")}

\item{beta.thr}{numeric, beta difference threshold value. Default = 0.3.}

\item{overlap.param.thr}{numeric, threshold value for selected parameter to filter methylations overlapping the selected features. Default = 100}

\item{param.type}{character, threshold parameter to filter methylations overlapping the selected features. Accetped c("dmr.length", "overlap.length", "overlap.percentage"). Default = "overlap.length".}

\item{line.col}{character, color of lines at x=0, y=0. Defalt = "lightgray"}

\item{lmfit.col1}{character, color of linear model line 1 or for simple plot. Defalt = "red"}

\item{lmfit.col2}{character, color of linear model line 2. Defalt = "green"}

\item{pal}{character, color palette. hcl.pals() to show available. Default = "RdGy"}

\item{plot.type}{character, compute or not different linear models for upregulated and downregulated genes. Accepted: "simple" or "splitted". Default = "splitted".}

\item{show.text}{logical, indicating if print gene names in the final plot. Accepted values: TRUE or FALSE. Default = FALSE.}

\item{cor.type}{character, correlation method. Available "pearson", "kendall" or "spearman" correlation, Default = "pearson"}

\item{return.table}{logical, TRUE return a data.frame instead a plot. Default = FALSE}
}
\value{
plot of correlation or (if return.table = TRUE) a data.frame of genes expression associated to beta methylation values of enhancers.
}
\description{
Compute and plot correlation between expression and methylation
}
