% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotateDMRs.R
\name{annotateDMRs}
\alias{annotateDMRs}
\title{Annotates the the differentially methylated regions.}
\usage{
annotateDMRs(
  DMRsRanges,
  prom.length = 1500,
  head.length = 1500,
  longest.trx = TRUE,
  annotation = "ensembl",
  hg = "hg19",
  annotation.fast = TRUE,
  thr.beta = 0.3,
  thr.cgis = 0.4,
  col.betadiff = 4,
  col.beta1 = NULL,
  col.beta2 = NULL
)
}
\arguments{
\item{DMRsRanges}{data.frame, the DMRs ranges, it must have the following columns: chr, start, end, beta diff. Other columns will be stored in the resulting output under the column other.}

\item{prom.length}{numeric, length of promoters. Default = 1500}

\item{head.length}{numeric, length of the first part of the gene, named head, starting from the TSS. If longer than the gene, the entire txs will be considered as head. Default = 1500}

\item{longest.trx}{logical, option to use the longest transcript to represent the gene}

\item{annotation}{character, database to use for transcripts mapping. Available "ensembl" or "ucsc". Default ="ensembl"}

\item{hg}{character, Available "hg19", "hg38". Genome assembly version. Default = "hg19"}

\item{annotation.fast}{logical, compute 1:1 mapping or 1:many - many:1 - many:many mapping. Default = TRUE}

\item{thr.beta}{numeric, beta difference threshold to consider methylations. Default = 0.3}

\item{thr.cgis}{numeric, length, in percentage, of methylated CGIs in order to be considered altered. Default = 0.4}

\item{col.betadiff}{nuemric, column position for beta diff. in input table. Default = 4}

\item{col.beta1}{numeric, column position for first sample beta values in input table}

\item{col.beta2}{numeric, column position for second sample beta values in input table}
}
\value{
list, features overlapped by annotated DMRs
}
\description{
Maps DMRs on genes returning a list for each features.
}
