% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotateEnhancers.R
\name{annotateEnhancers}
\alias{annotateEnhancers}
\title{Query database to find enhancer.}
\usage{
annotateEnhancers(
  DMRsRanges,
  hg = "hg19",
  thr.beta = 0.3,
  overlap.param.thr = 40,
  param.type = "overlap.percentage",
  score.modifier = 0.5,
  col.betadiff = 4
)
}
\arguments{
\item{DMRsRanges}{the DMRs ranges, it must have the following columns: chr, start, end, beta diff. Other columns will be stored in the resulting output under the column other.}

\item{hg}{character, Available "hg19", "hg38". Genome assembly version. Default = "hg19"}

\item{thr.beta}{numeric, beta difference threshold to consider methylations. Default = 0.3}

\item{overlap.param.thr}{numeric, threshold value for selected parameter to filter methylations overlapping the selected features. Default = 100}

\item{param.type}{character, threshold parameter to filter methylations overlapping the selected features. Accetped c("dmr.length", "overlap.length", "overlap.percentage"). Default = "overlap.length".}

\item{score.modifier}{numeric, value between 0-1. It specifies how the final score is computed by assigning different weights to the methylation charactersistics of enhacners or to genes already involved in pathologies. By increasing this value to 1, resulting scores will be focused on discovering segments affecting gene expression. A value equal to 0 will focus the results on enahcners involving genes associated to pathologies, not considering the effect of methylation. Default = 0.5}

\item{col.betadiff}{numeric, column position for beta diff. in input table. Default = 4}
}
\value{
a vector of presence or a data.frame
}
\description{
Query database of annotations results list to enhancer database in order to associate differentially methylated segments to genes.
}
