% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatDMRsInput.R
\name{formatDMRsInput}
\alias{formatDMRsInput}
\title{Convert input table to proper format.}
\usage{
formatDMRsInput(
  tableIn,
  thr.beta,
  col.betadiff = 4,
  col.beta1 = NULL,
  col.beta2 = NULL
)
}
\arguments{
\item{thr.beta}{numeric, beta difference threshold to consider methylations. Default = 0.3}

\item{col.betadiff}{numeric, column position for beta diff. in input table. Default = 4}

\item{col.beta1}{numeric, column position for first sample beta values in input table}

\item{col.beta2}{numeric, column position for second sample beta values in input table}

\item{DMRsRanges}{the DMRs ranges, it must have the following columns: chr, start, end, beta diff. Other columns will be stored in the resulting output under the column other.}
}
\value{
data.frame of DMR ranges
}
\description{
Convert input table to proper format. The firts three column of the input table must have chr, start, end coordinates.
}
