% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genesToNCG.R
\name{genesToNCG}
\alias{genesToNCG}
\title{Find genes annotated in NCG.}
\usage{
genesToNCG(annotatedDMRs, ncg, return.table = FALSE)
}
\arguments{
\item{annotatedDMRs}{anotated DMRs list resultingfrom annotateDMRs() or scoreAnnotatedDMRs()}

\item{ncg}{the NCG gene vectors}

\item{return.table}{logical, option TRUE return a table instead a vector of presences. Default = TRUE}
}
\value{
data.frame or vector of presences
}
\description{
Assess the presence of genes in results from annotateDMRs() in NCG database.
}
