% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDMRs.R
\name{plotDMRs}
\alias{plotDMRs}
\title{Converts annotated DMRs in a plot}
\usage{
plotDMRs(
  annotatedDMRs,
  symbol,
  annotation = "ensembl",
  hg = "hg19",
  beta1.name = NULL,
  beta2.name = NULL,
  beta.colors = c("red", "navy"),
  blackandwhite = FALSE,
  show.all.transcripts = FALSE,
  prom.width = 1500,
  path = NULL,
  coord.zoom = NULL,
  smartzoom = TRUE,
  height.pdf = 9,
  width.pdf = 16
)
}
\arguments{
\item{annotatedDMRs}{anotated DMRs list resultingfrom annotateDMRs() or scoreAnnotatedDMRs()}

\item{symbol}{character, gene symbol to plot.}

\item{annotation}{character, "ensembl" or "ucsc". Annotation used to track the plot. Default = "esembl".}

\item{hg}{character, "hg19" or "hg38". Genome Assembly version. Default = "hg19".}

\item{beta1.name}{character, if unsued beta difference is plotted. character string identifying beta value of first sample in "other" column in results from annotateDMRs() or indentifying colname in input table used in annotateDMRs()}

\item{beta2.name}{character, it identifies beta value of second sample in "other" column in results from annotateDMRs() or it identifies colname in input table used in annotateDMRs()}

\item{beta.colors}{character vectors, colors of tracks for the first and the second bvalue, respectivetely. Default is c("red", "navy"). If beta diff is plotted, only the first element of vector is considered.}

\item{blackandwhite}{logical, it allows to get all the plot in greyscale. Default = FALSE.}

\item{show.all.transcripts}{logical, if TRUE all transcripts of genes are tracked, if FALSE only the longest transcript is tracked. Default = FALSE.}

\item{prom.width}{integer, promoter lenght. Default = 1500.}

\item{path}{logical, path where the plot is saved in a pdf file. If NULL the plot is not saved. Default = NULL.}

\item{coord.zoom}{numeric vectors, coordinates of zoom region. If NULL the plot is not zoomed. Default = NULL.}

\item{smartzoom}{logical, automatic zoom on the methylated region. Default = TRUE.}

\item{height.pdf}{integer, hight pdf file. Default = 9.}

\item{width.pdf}{integer, width pdf file. Default = 16.}
}
\value{
Plot of the beta value(s) mapped on transcript(s).
}
\description{
This function allows to track in a plot the beta value of a methylated segment mapped on a transcript of the human genome
}
