% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDMRs2Enrichr.R
\name{plotDMRs2Enrichr}
\alias{plotDMRs2Enrichr}
\title{Plot EnrichR Results}
\usage{
plotDMRs2Enrichr(
  enrichr.results,
  annotatedDMRs,
  stat = "P.value",
  n = 25,
  plot.type = "barplot",
  pal.col = "Dynamic",
  col.hyper = "#ff0000",
  col.hypo = "#00b3ff",
  thrs = c(0.01, 0.05),
  thrs.cols = c("green", "yellow")
)
}
\arguments{
\item{enrichr.results}{data.frame resulting list from annotatedDMTs2Enrichr().}

\item{annotatedDMRs}{anotated DMRs list resultingfrom annotateDMRs() or scoreAnnotatedDMRs()}

\item{stat}{character, statistics to visualize. Accepted "P.value", "Adjusted.P.value" or "Overlap". Default = 'P.value'.}

\item{n}{numeric, value indicating the number of enrichment to plot, starting from the most enriched, to visualize. Default = 25.}

\item{plot.type}{character, compute or not different linear models for upregulated and downregulated genes. Accepted: "simple" or "splitted". Default = "splitted".}

\item{pal.col}{character, the palette color for plot.type = 'lollipop'. Must be one from hcl.pals(). Default = 'Dynamic'.}

\item{col.hyper}{character, the color representing hyper-methylated genes. Default = 'Grey70'.}

\item{col.hypo}{character, the color representing hyper-methylated genes. Default = 'Grey30'.}

\item{thrs}{numeric vectors, statistical thresholds to plot. Default = c(0.01, 0.05)}

\item{thrs.cols}{characters vector, the colors to use for thresholds. Default = c('green', "yellow").}
}
\value{
plot of enrichR results
}
\description{
Visualize enrichment results and the contributes of hyper-metyhlated and hypo-methylated genes.
}
