% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMethylationOverview.R
\name{plotMethylationOverview}
\alias{plotMethylationOverview}
\title{Plot a simple distribution of methylations.}
\usage{
plotMethylationOverview(annotatedDMRs, plot.type, palette)
}
\arguments{
\item{annotatedDMRs}{anotated DMRs list resultingfrom annotateDMRs() or scoreAnnotatedDMRs()}

\item{plot.type}{character, compute or not different linear models for upregulated and downregulated genes. Accepted: "simple" or "splitted". Default = "splitted".}

\item{palette}{character, color palette of plot. It must be one resulting from hcl.pals()}
}
\value{
plot of enrichR results
}
\description{
Plot distribution of beta values on chromosomes, this function has mainly an internal utility and is design for shinyApp version of package
}
