% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTFs2Exprs.R
\name{plotTFs2Exprs}
\alias{plotTFs2Exprs}
\title{Plot the beta value of TF and related target genes expression}
\usage{
plotTFs2Exprs(
  associatedTFs2Expr,
  symbol,
  col.meth = "#8e0000",
  pals.bars = "Cold"
)
}
\arguments{
\item{associatedTFs2Expr}{data.frame, results from associateTFs2Exprs().}

\item{symbol}{character, feature to correlate. Must be from names of resulting list from annotateDMRs. Additional feature names can be first exons (exons1) or first intron (intron1). To use more than one feature use c().}

\item{col.meth}{character, color for beta value. Defaut = "#8e0000" (red)}

\item{pals.bars}{character, palette for gene expresion. hcl.pals() to show available. Default = "Cold"}
}
\value{
plot showing methylation levels of TF and expresion of target genes
}
\description{
Barplot of the beta value of TF and the expression of the target genes
}
