% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryDatabase.R
\name{queryDatabase}
\alias{queryDatabase}
\title{Query database to find pathologic genes.}
\usage{
queryDatabase(
  DMRsRanges,
  db,
  return.table = TRUE,
  hold.columns,
  is.genomic.ranges = FALSE,
  thr = 0
)
}
\arguments{
\item{DMRsRanges}{DMRs ranges}

\item{db}{character, database to query}

\item{return.table}{logical, optition to return a table instead a vector of presence. Default = TRUE}

\item{hold.columns}{numeric vectors, column positions to hold when return.table=TRUE}

\item{is.genomic.ranges}{logical, specifies if database is a GenomicRange objet or a data.frame. Default = FALSE}

\item{thr}{numeric, threshold for beta difference values. Default = 0}
}
\value{
a vector of presence or a data.frame
}
\description{
Query elements of resulting annotations list to specified database in order to find pathogenic genes.
}
