% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreAnnotatedDMRs.R
\name{scoreAnnotatedDMRs}
\alias{scoreAnnotatedDMRs}
\title{Score the annotated methylation segments.}
\usage{
scoreAnnotatedDMRs(
  annotatedDMRs,
  active.features = c("promoters", "heads"),
  score.modifier = 0.5
)
}
\arguments{
\item{annotatedDMRs}{anotated DMRs list resultingfrom annotateDMRs()}

\item{active.features}{character vectors, containing features to correlate. Must be from names of resulting list from annotateDMRs. Additional feature names can be first exons (exons1) or first intron (intron1). To use more than one feature use c(). Default = c("promoters", "heads")}

\item{score.modifier}{numeric, value between 0-1. It specifies how the final score is computed by assigning different weights to the methylation charactersistics of enhacners or to genes already involved in pathologies. By increasing this value to 1, resulting scores will be focused on discovering segments affecting gene expression. A value equal to 0 will focus the results on enahcners involving genes associated to pathologies, not considering the effect of methylation. Default = 0.5}
}
\value{
data.frame of annotated DMRs with assigned scores
}
\description{
Assigns a score to annotated methylated segments resulting from annotateDMRs() function
}
