% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfs2Enrichr.R
\name{tfs2Enrichr}
\alias{tfs2Enrichr}
\title{Query different databases to find enriched proceses.}
\usage{
tfs2Enrichr(
  associatedTFs2Expr,
  logfc.thr = 1,
  stat.filter = "P.value",
  stat.thr = 0.01,
  db = NULL
)
}
\arguments{
\item{associatedTFs2Expr}{data.frame. Corresponding to resulting data.frame from associateTFs2Exprs().}

\item{logfc.thr}{numeric value indicating logFC threshold. Default = 1.}

\item{stat.filter}{character indicating which type of statistics use for filtering results. Accepted values: 'P.value' or 'Adjusted.P.value'. Default 'P.value'.}

\item{stat.thr}{numeric value indicating the threshold to use for selcted statistical test. Default 0.01.}

\item{db}{vector of characters indicating DBs to query. Default c("ClinVar_2019", "OMIM_Disease", "Elsevier_Pathway_Collection", "MSigDB_Hallmark_2020", "MSigDB_Oncogenic_Signatures", "GO_Biological_Process_2018", "Human_Phenotype_Ontology", "KEGG_2016", "NCI-Nature_2016", "Panther_2016", "Reactome_2016", "WikiPathways_2019_Human")}
}
\value{
data.frame with enriched processes
}
\description{
Query TF's targeted genes in order to find enriched proceses.
}
